Protect readline calls in try blocks
Sometimes the readline encounters a non-UTF8
character, which causes the program to crash.
Protect this with try calls and catch the exception.
In this case we will miss the relevant information for
that particular scan point but the scan will continue
as normal.

Better would be to figure out the correct encoding (while still
keeping the try block just in case).