Plot BER vs TAP0 summary plot
Add a plotting function summarizing the information across
may scans, showing the Bit Error Rate as a function of TAP0.

Data are shown as line plots in (TAP0, BER) as (X,Y) with
the positions separated out as different columns, different modules
shown in different colours, and different chips with different
linestyles.

Currently the results look like a 5-year old was let loose with
a crayon.

May need to consider what kind of data selection/grooming is used
for this plot.