Adding test and timing command line options
If the `--test` command line option is used, then results won't be logged
in the database. Temperature and VDDD/VDDA checking will also not be done
before each scan.

the length of each scan point in seconds can be set with the command
line option `-t` followed by the number of seconds per scan point.