WIP: First attempt at script for archiving
Now that we have the tuning database, also working on
an archive script to archive the good runs. This is a
first simple shot at that. Not all features are implemented,
in particular the plot copying is not done yet, as well as some
other minor things.

Basic way it works is it flags runs as archived in the tuning db,
and also copies results files over to a subdirectory of the archive
folder, one directory per run.